% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_cases}
\alias{example_cases}
\title{Example Case Data}
\format{
A data frame with 13 rows and 6 columns:
\describe{
\item{id}{Case id}
\item{adm3}{Administrative district}
\item{outcome}{Exit status}
\item{date_symptoms}{Date of symptom onset}
\item{date_hospital}{Date of hospitalization}
\item{date_outcome}{Date of case outcome (discharged or died)}
}
}
\usage{
example_cases
}
\description{
An example (minimal) linelist of individual case data. This data is fake and is not intended
to represent any particular pathogen.
}
\keyword{datasets}
