% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_pairs}
\alias{example_pairs}
\title{Example Infector-Infectee Pair Data}
\format{
A data frame with 13 rows and 2 columns:
\describe{
\item{id}{Case id}
\item{infector_id}{Infector id}
}
}
\usage{
example_pairs
}
\description{
An example (minimal) dataset of infector-infectee pairs corresponding to \code{casees}.
}
\keyword{datasets}
