% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cases.R
\name{map_case}
\alias{map_case}
\title{Map Secondaries of a Case}
\usage{
map_case(
  case_id,
  order_id,
  cases,
  var_id = "id",
  var_infector = "infector_id",
  verbose = FALSE
)
}
\arguments{
\item{case_id}{\code{chr} Case to map.}

\item{order_id}{\code{int} Case order in dataset.}

\item{cases}{\code{df} Data frame of case data.}

\item{var_id}{\code{chr} Column name for case id.}

\item{var_infector}{\code{chr} Column name for infector id.}

\item{verbose}{\code{bool} Whether to print output on cases being mapped.}
}
\description{
Internal function that calculates visual order for the secondaries of a single case.
}
