% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cases.R
\name{map_cases}
\alias{map_cases}
\title{Map Order of Cases}
\usage{
map_cases(
  cases,
  pairs,
  var_id = "id",
  var_infector = "infector_id",
  var_order = "date_symptoms",
  verbose = FALSE
)
}
\arguments{
\item{cases}{\code{df} Data frame of case data.}

\item{pairs}{\code{df} Data frame with infector infectee pairs.}

\item{var_id}{\code{chr} Column name for case id.}

\item{var_infector}{\code{chr} Column name for infector id.}

\item{var_order}{\code{chr} Column name for the variable used to visually order cases.}

\item{verbose}{\code{bool} Whether to print output on cases being mapped.}
}
\description{
Internal function that calculates visual order to position cases in linear and excel
output.
}
