% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chains_linear.R
\name{plot_chains_linear}
\alias{plot_chains_linear}
\title{Plot Linear Chains (ggplot2)}
\usage{
plot_chains_linear(
  cases,
  height = 5,
  gap = height/2,
  offset = 1,
  var_id = "id",
  var_infector = "infector_id",
  var_order = var_symptoms,
  var_outcome_status = "outcome",
  var_symptoms = "date_symptoms",
  var_transition = "date_hospital",
  var_outcome = "date_outcome",
  fill_infected = "#EEEEEE",
  fills_outcome = c("#000000", "#819CBB"),
  color_links = "#000000",
  date_breaks = "1 week",
  verbose = FALSE
)
}
\arguments{
\item{cases}{\code{df} Dataframe of case information. This dataset should contain information on
outcome, onset date, transition (hospitalization) date, and outcome date, and infector.}

\item{height}{\code{num} Height of bars representing cases.}

\item{gap}{\code{num} Gap between bars.}

\item{offset}{\code{num} Offset between the beginning of a bar and the link connecting it to any
secondary cases.}

\item{var_id}{\code{chr} Column name for case id.}

\item{var_infector}{\code{chr} Column name for infector id.}

\item{var_order}{\code{chr} Column name for the variable used to visually order cases.}

\item{var_outcome_status}{\code{chr} Column name for outcome status.}

\item{var_symptoms}{\code{chr} Column name for symptom onset date.}

\item{var_transition}{\code{chr} Column name for transition (hospitalization) date.}

\item{var_outcome}{\code{chr} Colun name for outcome date.}

\item{fill_infected}{\code{chr} Color to use for initial infected period of the bar.}

\item{fills_outcome}{\verb{chr vct} Colors to use for the second part of the bar, indicating outcome.}

\item{color_links}{\code{chr} Color for lines linking primary and secondary cases.}

\item{date_breaks}{\code{chr} Breaks to use for date axis.}

\item{verbose}{\code{bool} Whether to print output on cases being mapped.}
}
\description{
Plots transmission trees with a linear layout, showing infections and case
progression over time. By default cases are ordered by date of symptom onset, starting from
top left to bottom right. \strong{Note} the function expects case data to have a column indicating the
id of the infector, if your pairs are stored in a separate dataset, you must join them first.
}
\examples{
cases <- left_join(example_cases, example_pairs) # add infector information
plot_chains_linear(cases)

}
